Move form H2 in-memory tests to Docker Testcontainer tests
H2 does not support some SQL function such as date_trunc.
As such, I could not implement a UNIQUE constraint on the event_template database table.
Thus I decided to move away from H2. This slow down a little bit the integration tests but it is still acceptable.
