+++ src/test/resources/application.properties
@@ -14,31 +14,28 @@
 # limitations under the License.
 #
 
-# Configuration file
-# key = value
+# Configuration file for integration testing
 
 quarkus.http.test-port=8092
 
-#DATA_BASE Postgresql
-quarkus.datasource.db-kind=h2
-quarkus.datasource.jdbc.url=jdbc:h2:mem:timekeeper;MODE=PostgreSQL
-quarkus.datasource.jdbc.driver=org.h2.Driver
+# DATA_BASE Postgresql
+# Switched from H2 to Docker PGSQL as H2 does not support some SQL function such as date_truncate
+
+# initializes container for driver initialization
+%test.quarkus.datasource.driver=org.testcontainers.jdbc.ContainerDatabaseDriver
+# dialect must be set explicitly
+%test.quarkus.hibernate-orm.dialect=org.hibernate.dialect.PostgreSQL9Dialect
+# Testcontainers JDBC URL
+%test.quarkus.datasource.url=jdbc:tc:postgresql:latest:///db
+%test.quarkus.datasource.username=demo
+%test.quarkus.datasource.password=demo
+
 #Hibernate
 quarkus.hibernate-orm.database.generation=none
 
 #Flyway
 quarkus.flyway.migrate-at-start=true
 quarkus.flyway.locations=db/migration
-# Flyway optional config properties
-# quarkus.flyway.baseline-on-migrate=true
-# quarkus.flyway.baseline-version=1.0.0
-# quarkus.flyway.baseline-description=Initial version
-# quarkus.flyway.connect-retries=10
-# quarkus.flyway.schemas=TEST_SCHEMA
-# quarkus.flyway.table=flyway_quarkus_history
-# quarkus.flyway.sql-migration-prefix=X
-# quarkus.flyway.repeatable-sql-migration-prefix=K
-# security configurations
 
 quarkus.oidc.auth-server-url=http://localhost:8180/auth/realms/Timekeeper
 quarkus.oidc.client-id=timekeeper-quarkus-backend
